remove_apache_http_client - move folder client over to retrofit
