+++ src/main/java/ploiu/service/FileService.java
@@ -11,16 +11,17 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.ResponseBody;
 import org.jetbrains.annotations.Nullable;
+import ploiu.client.DeprecatedFileClient;
 import ploiu.client.FileClient;
-import ploiu.client.RetrofitFileClient;
-import ploiu.model.*;
+import ploiu.model.CreateFileRequest;
+import ploiu.model.FileApi;
+import ploiu.model.FileSearch;
+import ploiu.model.UpdateFileRequest;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.nio.file.Files;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static ploiu.Constants.CACHE_DIR;
@@ -29,8 +30,8 @@ import static ploiu.Constants.LIST_IMAGE_SIZE;
 @Slf4j
 @RequiredArgsConstructor(onConstructor_ = @Inject)
 public class FileService {
-    private final FileClient fileClient;
-    private final RetrofitFileClient retrofitFileClient;
+    private final DeprecatedFileClient fileClient;
+    private final FileClient retrofitFileClient;
 
     public Single<File> saveAndGetFile(FileApi fileApi, @Nullable File directory) {
         return Single.fromCallable(() -> {
@@ -73,31 +74,6 @@ public class FileService {
 
     public Single<FileApi> createFile(CreateFileRequest request) {
         return Single.just(request).observeOn(Schedulers.io()).subscribeOn(Schedulers.io()).flatMap(fileClient::createFile);
-    }
-
-    /**
-     * returns files icon images for each file in the passed folder.
-     * <p>
-     * this method will attempt to retrieve cached icons first, and it pairs them with the corresponding file id
-     * (to prevent re-uploaded files with the same name but different contents from having the same preview).
-     *
-     * @param folder
-     * @return
-     */
-    public Single<Map<Long, Image>> getFilePreviewsForFolder(FolderApi folder) {
-        return Observable.just(folder)
-                .observeOn(Schedulers.io())
-                .map(FolderApi::id)
-                .flatMapSingle(retrofitFileClient::getPreviewsForFolder)
-                .map(map -> {
-                    Map<Long, Image> images = new HashMap<>();
-                    for (var entry : map.entrySet()) {
-                        var image = new Image(new ByteArrayInputStream(entry.getValue()), LIST_IMAGE_SIZE, LIST_IMAGE_SIZE, true, true);
-                        images.put(entry.getKey(), image);
-                    }
-                    return images;
-                })
-                .single(Map.of());
     }
 
     public Maybe<Image> getFilePreview(Long id) {
