+++ src/test/java/ploiu/service/DragNDropServiceTests.java
@@ -9,7 +9,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import ploiu.TestHelper;
-import ploiu.client.FolderClient;
 import ploiu.model.FileApi;
 import ploiu.model.FolderApi;
 import ploiu.model.FolderRequest;
@@ -30,7 +29,7 @@ class DragNDropServiceTests {
     static FolderApi rootApi = new FolderApi(0, 0, "root", null, List.of(), List.of(), List.of());
 
     @Mock
-    FolderClient folderClient;
+    FolderService folderService;
     @Mock
     FileService fileService;
 
@@ -53,10 +52,10 @@ class DragNDropServiceTests {
     @Test
     void testUploadFoldersSingleFolder() throws IOException {
         var dir = helper.createDir("top");
-        when(folderClient.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
         service.uploadFolders(List.of(dir), rootApi).toList().blockingGet();
-        verify(folderClient).createFolder(eq(new FolderRequest(Optional.empty(), 0, "top", List.of())));
-        verifyNoMoreInteractions(folderClient);
+        verify(folderService).createFolder(eq(new FolderRequest(Optional.empty(), 0, "top", List.of())));
+        verifyNoMoreInteractions(folderService);
         verifyNoInteractions(fileService);
     }
 
@@ -65,10 +64,10 @@ class DragNDropServiceTests {
         var first = helper.createDir("top");
         var second = helper.createDir("middle");
         var third = helper.createDir("bottom");
-        when(folderClient.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
         service.uploadFolders(List.of(first, second, third), rootApi).toList().blockingGet();
-        verify(folderClient, times(3)).createFolder(argThat(arg -> arg.id().isEmpty() && arg.parentId() == 0));
-        verifyNoMoreInteractions(folderClient);
+        verify(folderService, times(3)).createFolder(argThat(arg -> arg.id().isEmpty() && arg.parentId() == 0));
+        verifyNoMoreInteractions(folderService);
         verifyNoInteractions(fileService);
     }
 
@@ -76,12 +75,12 @@ class DragNDropServiceTests {
     void testUploadFoldersNestedFolders() throws IOException {
         var top = helper.createDir("top");
         helper.createDir("top/middle/bottom");
-        when(folderClient.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(2, 1, "middle", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(2, 1, "middle", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
         service.uploadFolders(List.of(top), rootApi).toList().blockingGet();
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 0 && it.name().equals("top")));
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 1 && it.name().equals("middle")));
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 2 && it.name().equals("bottom")));
-        verifyNoMoreInteractions(folderClient);
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 0 && it.name().equals("top")));
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 1 && it.name().equals("middle")));
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 2 && it.name().equals("bottom")));
+        verifyNoMoreInteractions(folderService);
         verifyNoInteractions(fileService);
     }
 
@@ -91,15 +90,15 @@ class DragNDropServiceTests {
         helper.createFile("top/first.txt");
         helper.createFile("top/second.txt");
         helper.createFile("top/third.txt");
-        when(folderClient.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "", null, List.of(), List.of(), List.of())));
         when(fileService.createFile(any())).thenReturn(Single.just(new FileApi(0, "", List.of(), null)));
         service.uploadFolders(List.of(dir), rootApi).toList().blockingGet();
-        verify(folderClient).createFolder(any());
+        verify(folderService).createFolder(any());
         verify(fileService).createFile(argThat(it -> it.file().getName().equals("first.txt") && it.folderId() == 1));
         verify(fileService).createFile(argThat(it -> it.file().getName().equals("second.txt") && it.folderId() == 1));
         verify(fileService).createFile(argThat(it -> it.file().getName().equals("third.txt") && it.folderId() == 1));
         verifyNoMoreInteractions(fileService);
-        verifyNoMoreInteractions(folderClient);
+        verifyNoMoreInteractions(folderService);
     }
 
     @Test
@@ -136,23 +135,23 @@ class DragNDropServiceTests {
         helper.createFile("top/middle/bottom/bottomFirst.txt");
         helper.createFile("top/middle/bottom/bottomSecond.txt");
 
-        when(folderClient.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(any())).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of()))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
 
-        when(folderClient.createFolder(eq(new FolderRequest(Optional.empty(), 0, "top", List.of())))).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of())));
-        when(folderClient.createFolder(eq(new FolderRequest(Optional.empty(), 1, "middle", List.of())))).thenReturn(Single.just(new FolderApi(2, 1, "middle", null, List.of(), List.of(), List.of())));
-        when(folderClient.createFolder(eq(new FolderRequest(Optional.empty(), 1, "middle2", List.of())))).thenReturn(Single.just(new FolderApi(4, 1, "middle2", null, List.of(), List.of(), List.of())));
-        when(folderClient.createFolder(eq(new FolderRequest(Optional.empty(), 2, "bottom", List.of())))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(eq(new FolderRequest(Optional.empty(), 0, "top", List.of())))).thenReturn(Single.just(new FolderApi(1, 0, "top", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(eq(new FolderRequest(Optional.empty(), 1, "middle", List.of())))).thenReturn(Single.just(new FolderApi(2, 1, "middle", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(eq(new FolderRequest(Optional.empty(), 1, "middle2", List.of())))).thenReturn(Single.just(new FolderApi(4, 1, "middle2", null, List.of(), List.of(), List.of())));
+        when(folderService.createFolder(eq(new FolderRequest(Optional.empty(), 2, "bottom", List.of())))).thenReturn(Single.just(new FolderApi(3, 2, "bottom", null, List.of(), List.of(), List.of())));
 
         when(fileService.createFile(any())).thenReturn(Single.just(new FileApi(0, "", List.of(), null)));
 
         service.uploadFolders(List.of(top), rootApi).toList().blockingGet();
 
         /// FOLDERS
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 0 && "top".equals(it.name())));
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 1 && "middle".equals(it.name())));
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 2 && "bottom".equals(it.name())));
-        verify(folderClient).createFolder(argThat(it -> it.parentId() == 1 && "middle2".equals(it.name())));
-        verifyNoMoreInteractions(folderClient);
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 0 && "top".equals(it.name())));
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 1 && "middle".equals(it.name())));
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 2 && "bottom".equals(it.name())));
+        verify(folderService).createFolder(argThat(it -> it.parentId() == 1 && "middle2".equals(it.name())));
+        verifyNoMoreInteractions(folderService);
         /// FILES
         // top
         verify(fileService).createFile(argThat(it -> it.folderId() == 1 && "topFirst.txt".equals(it.file().getName())));
